# docs/contracts/openapi/admin.openapi.yaml
openapi: 3.1.0
info:
  title: OnMaru Admin API
  version: 1.0.0-draft
  description: >-
    관리자 콘솔용 인증, 대시보드, 후기/신고 검수, 사용자 제재, 큐레이션,
    데이터 파이프라인 운영 계약. 모든 변경은 감사 로그와 멱등성 키를 사용한다.
servers:
  - url: /api/v1
...
paths:
...
  /admin/curations:
    get:
      tags: [AdminCurations]
      operationId: listAdminCurations
      summary: 큐레이션 대상 목록 조회
      security: [{adminBearer: []}]
      parameters:
        - {name: category, in: query, required: true, schema: {type: string, enum: [VILLAGE, STAY, ROUTE]}}
        - {name: included, in: query, schema: {type: string, enum: [ALL, INCLUDED, EXCLUDED], default: ALL}}
        - {name: query, in: query, schema: {type: string, maxLength: 120}}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: 큐레이션 페이지
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurationPage'}
        '401': {$ref: '#/components/responses/Error'}
  /admin/curations/{placeId}:
    put:
      tags: [AdminCurations]
      operationId: upsertAdminCuration
      summary: 큐레이션 override 저장
      security: [{adminBearer: []}]
      parameters:
        - {name: placeId, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, included, badges]
              additionalProperties: false
              properties:
                category: {type: string, enum: [VILLAGE, STAY, ROUTE]}
                included: {type: boolean}
                badges: {type: array, items: {type: string, maxLength: 40}}
      responses:
        '200': {description: 저장된 curation override}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
  /admin/audit-logs:
    get:
      tags: [AdminAuditLogs]
      operationId: listAdminAuditLogs
      summary: 관리자 감사 로그 조회
      security: [{adminBearer: []}]
      parameters:
        - {name: actorAdminId, in: query, schema: {type: string, format: uuid}}
        - {name: resourceType, in: query, schema: {type: string, maxLength: 80}}
        - {name: resourceId, in: query, schema: {type: string, maxLength: 160}}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200': {description: 감사 로그 목록}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
  /admin/pipelines/{dataset}/status:
    get:
      tags: [AdminPipelines]
      operationId: getAdminPipelineStatus
      summary: 데이터 파이프라인 상태 조회
      security: [{adminBearer: []}]
      parameters: [{$ref: '#/components/parameters/Dataset'}]
      responses:
        '200':
          description: pipeline 상태
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PipelineStatus'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
  /admin/pipelines/{dataset}/runs:
    post:
      tags: [AdminPipelines]
      operationId: startAdminPipelineRun
      summary: 데이터 pipeline 수동 실행
      security: [{adminBearer: []}]
      parameters:
        - {$ref: '#/components/parameters/Dataset'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scope, reason]
              properties:
                scope: {type: string, enum: [FULL, VILLAGE, STAY, ROUTE]}
                reason: {type: string, minLength: 1, maxLength: 300}
      responses:
        '202':
          description: pipeline 실행 접수
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunAccepted'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminRefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-onmaru-admin-refresh
  parameters:
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 512}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, format: uuid}
    Dataset:
      name: dataset
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 80}
...
  headers:
    NoStore:
      schema: {type: string, const: no-store}
...
  responses:
    Error:
      description: 표준 API 오류
      headers:
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
...
  schemas:
...
    CurationPage:
      type: object
      required: [items, hasNext]
      properties:
        items: {type: array, items: {type: object}}
        nextCursor: {type: string, nullable: true}
        hasNext: {type: boolean}
    PipelineStatus:
      type: object
      required: [dataset, status]
      properties:
        dataset: {type: string}
        status: {type: string, enum: [IDLE, RUNNING, SUCCEEDED, FAILED, CANCELLED, MISSING]}
        lastSuccessAt: {type: string, format: date-time, nullable: true}
        failureCount: {type: integer, minimum: 0}
    RunAccepted:
      type: object
      required: [runId, status]
      properties:
        runId: {type: string, format: uuid}
        status: {type: string, enum: [QUEUED, RUNNING]}
...
    ApiError:
      type: object
      required: [schemaVersion, code, message, requestId]
      properties:
        schemaVersion: {type: string}
        code: {type: string}
        message: {type: string}
        requestId: {type: string, format: uuid}
        details: {type: object}